Use a stub implementation of musl global state when musl is embedded in a larger glibc program.
This commit hides all pthread internal layout information when musl is
being compiled as a layer atop glibc. It also changes the shape and
field names of some musl-internal global variables. This is done to more
easily catch unexpected dependencies on libc internal state.

Also, remove a stray reference to __aio_close from close(). It shows up
in the test output binaries and while it's benign, I don't like that
it's there.

FIXME: Currently, accesses to musl-internal locale data will trap; this
should be reevaluated and possibly layered on top of glibc.