setrlimit: Disable the SYS_setrlimit fallback when not using musl libc internals
The fallback uses __synccall to force all musl-known threads to run
SYS_setrlimit, but when layering musl on top of glibc, the data that
__synccall depends on is not available in an ABI-stable way, if it's
even available at all.

SYS_prlimit64 appears in the linux kernel source tree in v2.6.36, which
was released on 2010-10-20. No program that I personally care about will
simultaneously depend on prlimit64 and also run on kernel 2.6.36 or
older, so just disable the fallback entirely in glibc compatibility
mode.